Extracted PeerClient to external file
